+++ init.lua
@@ -4,5 +4,4 @@ require('colorscheme').setup()
 require('core')
 require('plugins')
 
-if vim.g.nvui then require('nvui') end
-
+if vim.g.neovide then require('neovide') end
