+++ lua/core/init.lua
@@ -1,4 +1,3 @@
 require('core.o')
 require('core.g')
 require('core.mappings')
-require('core.extra')
